+++ scripts/doc_snippets/get-started/install.py
@@ -1,0 +1,3 @@
+"""The reader's side of docs/get-started/install. See _common.py."""
+
+from _common import *  # noqa: F403